@@ -3,16 +3,20 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "axios": "^0.18.0",
-    "react": "^16.8.5",
-    "react-dom": "^16.8.5",
-    "react-particles-js": "^2.5.1",
-    "react-router-dom": "^5.0.0",
+    "@material-ui/core": "^4.11.0",
+    "@material-ui/icons": "^4.9.1",
+    "@material-ui/lab": "^4.0.0-alpha.56",
+    "axios": "^0.18.1",
+    "fontsource-roboto": "^3.0.3",
+    "react": "^16.14.0",
+    "react-dom": "^16.14.0",
+    "react-particles-js": "^2.7.1",
+    "react-router-dom": "^5.2.0",
     "react-scripts": "2.1.8"
   },
   "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
+    "start": "REACT_APP_API_HOST=http://localhost:8080 react-scripts start",
+    "build": "REACT_APP_API_HOST=https://shrturl-cc.herokuapp.com react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject"
   },
@@ -24,5 +28,10 @@
     "not dead",
     "not ie <= 11",
     "not op_mini all"
-  ]
+  ],
+  "devDependencies": {
+    "eslint": "^5.12.0",
+    "install": "^0.13.0",
+    "npm": "^6.14.8"
+  }
 }
