@@ -1,31 +1,32 @@
 import React from "react";
+import { Card, CardContent } from "@material-ui/core";
 import LinkListItem from "./LinkListItem";
 
-class LinkList extends React.Component {
-  shouldComponentUpdate(nextProps, nextState) {
-    if (nextProps.history.length !== this.props.history.length) {
-      return true;
-    }
-    return false;
+const LinkList = ({ history }) => {
+  if (history.length < 1) {
+    return null;
   }
 
-  render() {
-    if (this.props.history.length < 1) {
-      return null;
-    }
+  const list = history.slice(0, 5).map((link, i) => {
+    const { hash, longURL, shortURL, pageTitle } = link;
 
-    return this.props.history.slice(0, 5).map((link, i) => {
-      return (
-        <LinkListItem
-          id={link.hash + i}
-          key={link.hash + i}
-          longURL={link.longURL}
-          shortURL={link.shortURL}
-          pageTitle={link.pageTitle}
-        />
-      );
-    });
-  }
-}
+    return (
+      <LinkListItem
+        id={hash + i}
+        key={hash + i}
+        longURL={longURL}
+        shortURL={shortURL}
+        pageTitle={pageTitle}
+        hasDivider={history.length > 1 && i !== 0}
+      />
+    );
+  });
+
+  return (
+    <Card variant="outlined">
+      <CardContent>{list}</CardContent>
+    </Card>
+  );
+};
 
 export default LinkList;
