@@ -1,13 +1,24 @@
 import React from "react";
+import { makeStyles } from "@material-ui/core";
+import Alert from "@material-ui/lab/Alert";
 
-const WarningBanner = props => {
-  if (!props.warn) {
+const useStyles = makeStyles((theme) => ({
+  root: {
+    width: "100%",
+    position: "absolute",
+  },
+}));
+
+const WarningBanner = ({ warn, message }) => {
+  const classes = useStyles();
+
+  if (!warn) {
     return null;
   }
 
   return (
-    <div className="alert alert-danger fixed-top" style={{ borderRadius: "0" }}>
-      <small>{props.message}</small>
+    <div className={classes.root}>
+      <Alert severity="error">{message}</Alert>
     </div>
   );
 };
